Extract async sibling signature matcher
