+++ src/ILInspector.Analysis.Tests/MemberIdentityValueEqualityTests.cs
@@ -47,10 +47,10 @@ public class MemberIdentityValueEqualityTests
         Assert.Equal(left, right);
         Assert.Equal(left.GetHashCode(), right.GetHashCode());
         Assert.True(
-            LibraryBodyAnalysisBuilder
+            LibraryBodyAsyncSiblingSignatureMatcher
                 .AsyncSiblingTypesMatch(left, right));
         Assert.True(
-            LibraryBodyAnalysisBuilder
+            LibraryBodyAsyncSiblingSignatureMatcher
                 .AsyncSiblingTypeIdentity(left)
                 .Length < 10_000);
     }
@@ -131,9 +131,9 @@ public class MemberIdentityValueEqualityTests
 
         Assert.Equal(versionOne, versionTwo);
         Assert.NotEqual(
-            LibraryBodyAnalysisBuilder
+            LibraryBodyAsyncSiblingSignatureMatcher
                 .AsyncSiblingTypeIdentity(shared),
-            LibraryBodyAnalysisBuilder
+            LibraryBodyAsyncSiblingSignatureMatcher
                 .AsyncSiblingTypeIdentity(mixed));
     }
 
@@ -147,7 +147,7 @@ public class MemberIdentityValueEqualityTests
             value = TypeRef.GenericInstance(pair, [value, value]);
 
         var exception = Assert.Throws<BadImageFormatException>(
-            () => LibraryBodyAnalysisBuilder
+            () => LibraryBodyAsyncSiblingSignatureMatcher
                 .EnsureAsyncSiblingDisplayIsBounded(value));
         Assert.Contains(
             "output limit",
@@ -165,7 +165,7 @@ public class MemberIdentityValueEqualityTests
             TypeRef.CoreLib("System", "Void"),
             MemberKind.Method);
 
-        LibraryBodyAnalysisBuilder
+        LibraryBodyAsyncSiblingSignatureMatcher
             .EnsureAsyncSiblingDisplayIsBounded(member);
     }
 
@@ -181,7 +181,7 @@ public class MemberIdentityValueEqualityTests
             MemberKind.Method);
 
         Assert.Throws<BadImageFormatException>(
-            () => LibraryBodyAnalysisBuilder
+            () => LibraryBodyAsyncSiblingSignatureMatcher
                 .EnsureAsyncSiblingDisplayIsBounded(member));
     }
 
@@ -209,22 +209,22 @@ public class MemberIdentityValueEqualityTests
                 [1]));
 
         Assert.False(
-            LibraryBodyAnalysisBuilder.AsyncSiblingTypesMatch(
+            LibraryBodyAsyncSiblingSignatureMatcher.AsyncSiblingTypesMatch(
                 baseline,
                 differentSize));
         Assert.False(
-            LibraryBodyAnalysisBuilder.AsyncSiblingTypesMatch(
+            LibraryBodyAsyncSiblingSignatureMatcher.AsyncSiblingTypesMatch(
                 baseline,
                 differentLowerBound));
         Assert.NotEqual(
-            LibraryBodyAnalysisBuilder.AsyncSiblingTypeIdentity(
+            LibraryBodyAsyncSiblingSignatureMatcher.AsyncSiblingTypeIdentity(
                 baseline),
-            LibraryBodyAnalysisBuilder.AsyncSiblingTypeIdentity(
+            LibraryBodyAsyncSiblingSignatureMatcher.AsyncSiblingTypeIdentity(
                 differentSize));
         Assert.NotEqual(
-            LibraryBodyAnalysisBuilder.AsyncSiblingTypeIdentity(
+            LibraryBodyAsyncSiblingSignatureMatcher.AsyncSiblingTypeIdentity(
                 baseline),
-            LibraryBodyAnalysisBuilder.AsyncSiblingTypeIdentity(
+            LibraryBodyAsyncSiblingSignatureMatcher.AsyncSiblingTypeIdentity(
                 differentLowerBound));
     }
 
@@ -236,7 +236,7 @@ public class MemberIdentityValueEqualityTests
             rank: 1_000_000);
 
         Assert.Throws<BadImageFormatException>(
-            () => LibraryBodyAnalysisBuilder
+            () => LibraryBodyAsyncSiblingSignatureMatcher
                 .EnsureAsyncSiblingDisplayIsBounded(array));
     }
 
@@ -248,7 +248,7 @@ public class MemberIdentityValueEqualityTests
             array = TypeRef.MdArray(array, rank: 60_000);
 
         Assert.Throws<BadImageFormatException>(
-            () => LibraryBodyAnalysisBuilder
+            () => LibraryBodyAsyncSiblingSignatureMatcher
                 .EnsureAsyncSiblingDisplayIsBounded(array));
     }
 
@@ -262,7 +262,7 @@ public class MemberIdentityValueEqualityTests
             value = TypeRef.GenericInstance(pair, [value, value]);
 
         Assert.True(
-            LibraryBodyAnalysisBuilder
+            LibraryBodyAsyncSiblingSignatureMatcher
                 .IsSupportedAsyncSiblingType(value));
     }
 
@@ -299,7 +299,7 @@ public class MemberIdentityValueEqualityTests
                     name));
 
         Assert.False(
-            LibraryBodyAnalysisBuilder.AsyncSiblingTypesMatch(
+            LibraryBodyAsyncSiblingSignatureMatcher.AsyncSiblingTypesMatch(
                 Create(assembly, literal),
                 Create(assembly, nested)));
     }
